requirements.txt
# Requirements file for Content Recommendation Engine
# ================================================

# Core Azure SDK packages
azure-functions>=1.18.0
azure-storage-blob>=12.19.0
azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0
azure-monitor-opentelemetry>=1.2.0
azure-ai-ml>=1.12.0
azure-cognitiveservices-language-textanalytics>=5.3.0
azure-search-documents>=11.4.0

# Machine Learning packages
scikit-learn>=1.3.0
pandas>=2.1.0
numpy>=1.24.0
scipy>=1.11.0
mlflow>=2.8.0
joblib>=1.3.0

# Data processing
pyspark>=3.5.0
pyarrow>=14.0.0

# API and web framework
fastapi>=0.104.0
pydantic>=2.5.0
uvicorn>=0.24.0

# Monitoring and visualization
streamlit>=1.28.0
plotly>=5.17.0
dash>=2.14.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
requests>=2.31.0
aiohttp>=3.9.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
flake8>=6.1.0
bandit>=1.7.0
safety>=2.3.0

# Power BI integration
powerbiclient>=3.1.0
msal>=1.25.0

# Additional utilities for data generation
faker>=20.1.0
lorem>=0.1.1